Rename the function calling objects to llm objects
